add git and some other commands for easy terminal
here i add .bashrc file so many shorter commands for terminal for a better use and here i also add git commands as well as it shows your branch name. means on which branch you right now. so it is vary help-full for a git prescriptive. so download it and use it on your own terminal... if you get any issue then please tell me. i'll help you...